dre/requirements.txt
boto3>=1.34.0
snowflake-connector-python>=3.7.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
